+++ packages/segmented-button/segmented-button.js
@@ -6,11 +6,16 @@ import {
   provide,
   reactive,
   toRefs,
+  watch,
 } from 'vue';
 
 export default {
   name: 'mcw-segmented-button',
-  props: { singleSelect: Boolean, touch: Boolean },
+  props: {
+    singleSelect: Boolean,
+    touch: Boolean,
+    modelValue: { type: [Number, Array] },
+  },
   setup(props, { emit }) {
     const uiState = reactive({
       classes: {
@@ -75,6 +80,24 @@ export default {
       },
       notifySelectedChange: detail => {
         emit('change', detail);
+        if (Array.isArray(props.modelValue)) {
+          const { selected, index } = detail;
+          const idx = props.modelValue.indexOf(detail.index);
+          if (selected) {
+            idx < 1 &&
+              emit('update:modelValue', props.modelValue.concat(index));
+          } else {
+            idx > -1 &&
+              emit(
+                'update:modelValue',
+                props.modelValue
+                  .slice(0, idx)
+                  .concat(props.modelValue.slice(idx + 1)),
+              );
+          }
+        } else {
+          emit('update:modelValue', detail.index);
+        }
       },
     };
 
@@ -83,6 +106,38 @@ export default {
     onMounted(() => {
       foundation = new MDCSegmentedButtonFoundation(adapter);
       foundation.init();
+
+      if (props.singleSelect && props.modelValue !== void 0) {
+        foundation.selectSegment(props.modelValue);
+      }
+
+      watch(
+        () => props.modelValue,
+        nv => {
+          if (Array.isArray(nv)) {
+            const selectedSegments = mappedSegments.value
+              .filter(({ selected }) => selected)
+              .map(({ index }) => index);
+
+            // select the new ones
+            nv.forEach(v => {
+              if (selectedSegments.indexOf(v) < 0) {
+                foundation.selectSegment(v);
+              }
+            });
+
+            // unselect the ones that not there anymore
+            selectedSegments.forEach(v => {
+              if (nv.indexOf(v) < 0) {
+                foundation.unselectSegment(v);
+              }
+            });
+          } else {
+            foundation.selectSegment(nv);
+            foundation.handleSelected({ index: nv });
+          }
+        },
+      );
     });
 
     onBeforeUnmount(() => {