+++ packages/tooltip/tooltip.js
@@ -11,14 +11,19 @@ export default {
     const uiState = reactive({
       classes: {},
       styles: {},
+      surfaceStyle: {},
       rootAttrs: { 'aria-hidden': true },
       root: null,
+      isTooltipPersistent: false,
+      isTooltipRich: false,
     });
     let foundation;
     let anchorElem;
 
     const adapter = {
-      getAttribute: name => uiState.root.getAttribute(name),
+      getAttribute: name => {
+        return uiState.root.getAttribute(name);
+      },
       setAttribute: (attributeName, value) => {
         uiState.rootAttrs = { ...uiState.rootAttrs, [attributeName]: value };
       },
@@ -30,8 +35,19 @@ export default {
         const { [className]: removed, ...rest } = uiState.classes;
         uiState.classes = rest;
       },
+      getComputedStyleProperty: propertyName => {
+        return window
+          .getComputedStyle(uiState.root)
+          .getPropertyValue(propertyName);
+      },
       setStyleProperty: (property, value) =>
         (uiState.styles = { ...uiState.styles, [property]: value }),
+      setSurfaceStyleProperty: (propertyName, value) => {
+        uiState.surfaceStyle = {
+          ...uiState.surfaceStyle,
+          [propertyName]: value,
+        };
+      },
       getViewportWidth: () => window.innerWidth,
       getViewportHeight: () => window.innerHeight,
       getTooltipSize: () => {
@@ -43,6 +59,9 @@ export default {
 
       getAnchorBoundingRect: () => {
         return anchorElem ? anchorElem.getBoundingClientRect() : null;
+      },
+      getParentBoundingRect: () => {
+        return uiState.root.parentElement?.getBoundingClientRect() ?? null;
       },
       getAnchorAttribute: attr => {
         return anchorElem ? anchorElem.getAttribute(attr) : null;
@@ -152,10 +171,10 @@ export default {
       foundation = new MDCTooltipFoundation(adapter);
       foundation.init();
 
-      const isTooltipRich = foundation.getIsRich();
-      const isTooltipPersistent = foundation.getIsPersistent();
+      uiState.isTooltipRich = foundation.isRich();
+      uiState.isTooltipPersistent = foundation.isPersistent();
 
-      if (isTooltipRich && isTooltipPersistent) {
+      if (uiState.isTooltipRich && uiState.isTooltipPersistent) {
         anchorElem.addEventListener('click', handleClick);
       } else {
         anchorElem.addEventListener('mouseenter', handleMouseEnter);
@@ -172,11 +191,15 @@ export default {
 
     onBeforeUnmount(() => {
       if (anchorElem) {
-        anchorElem.removeEventListener('mouseenter', handleMouseEnter);
-        // TODO(b/157075286): Listening for a 'focus' event is too broad.
-        anchorElem.removeEventListener('focus', handleFocus);
-        anchorElem.removeEventListener('mouseleave', handleMouseLeave);
-        anchorElem.removeEventListener('blur', handleBlur);
+        if (uiState.isTooltipRich && uiState.isTooltipPersistent) {
+          anchorElem.removeEventListener('click', handleClick);
+        } else {
+          anchorElem.removeEventListener('mouseenter', handleMouseEnter);
+          // TODO(b/157075286): Listening for a 'focus' event is too broad.
+          anchorElem.removeEventListener('focus', handleFocus);
+          anchorElem.removeEventListener('mouseleave', handleMouseLeave);
+          anchorElem.removeEventListener('blur', handleBlur);
+        }
       }
 
       foundation?.destroy();